Added an icon for the aplication.
Changed the code so pyinstaller can add the icon to the executable.

To create exe with pyinstaller, first add ('./resources/*.png','resources'),('./resources/*.ico','resources') to datas[] in main.spec then run: pyinstaller main.spec --onefile

Adjusted the resource_path function so it is no longer necessary to change it when creating an executable with pyinstaller.