Remade the view elements so they can now be toggled
Fixed a problem with command history stack, where it would load the wrong command if the newly entered command was already at the top of the stack.

Adding files to the tree view now automatically expands the project where file is being added.

Imported projects are now automatically expanded in the tree view