When dropping multiple files, if some of them fail to be imported, they no longer trigger an error message individually. Instead they are all listed in one message.
Fixed an issue when project failed to be imported, and the project node was not created, but there was an attempt to acces it.

Reworked project name generation, so it now increments the number in the name of the project.

If file is dropped in the workspace, and the user chooses to create a project with the same name as the file, but if project with the same name as the file already exist, then a number will be incremented in the project name until it is available.